// gui/src/app/shared/interfaces/list-params.ts
// Copyright 2019, Pure Storage Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//    http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// tslint:disable:interface-name
// Available parameters for Collection.list()
export interface ListParams {
    pageStart?: number;
    pageSize?: number;
    sort?: SortParams | SortParams[];
    filter?: FilterParams;
}

export interface SortParams {
    key: string;
    order: 'asc' | 'desc';
}

export interface FilterParams {
    [key: string]: string | number;
}